feat: add internal error page to home
