+++ webapp/src/App.js
@@ -1,69 +1,22 @@
-import React, { useEffect, useState } from 'react'
+import React, { useState } from 'react'
 import PropTypes from 'prop-types'
 import Grid from '@material-ui/core/Grid'
-import { useLazyQuery, useQuery } from '@apollo/react-hooks'
-import { BrowserRouter, Route, Redirect, Switch, useHistory } from 'react-router-dom'
+import { MainContainer, TopBar, SideBar } from './containers'
+import { BrowserRouter, Route, Redirect, Switch } from 'react-router-dom'
 import { useCookies } from 'react-cookie'
 
 import routes from './routes'
-import { MainContainer, TopBar, SideBar } from './containers'
-import LocalBusinessStructuredData from './components/LocalBusinessStructuredData'
-import MedicalClinicStructuredData from './components/MedicalClinicStructuredData'
 import SplashIntro from './components/SplashIntro'
 import { useUser } from './context/user.context'
-import { GET_VALID_SPONSORS_QUERY, GET_VALID_LIFEBANKS_QUERY } from './gql'
 
 const App = ({ ual }) => {
-  const history = useHistory()
-  console.log(history)
-  const [validSponsors, setValidSponsors] = useState([])
-  const [validLifebanks, setValidLifebanks] = useState([])
   const [currentUser, { logout }] = useUser()
   const [cookies, setCookie] = useCookies(['splash'])
   const [sideBarPosition, setSideBarPosition] = useState(true)
-  const [internalError, setInternalError] = useState(false)
 
   const triggerSideBarPosition = () => {
     sideBarPosition ? setSideBarPosition(false) : setSideBarPosition(true)
   }
-
-  const { error: getValidSponsorError, data: validSponsorsData, refetch: getValidSponsors } = useQuery(GET_VALID_SPONSORS_QUERY, { fetchPolicy: 'cache-and-network' })
-
-  const getSponsors = async () => {
-    await getValidSponsors()
-
-    if (validSponsorsData && validSponsors.length === 0) setValidLifebanks(validSponsorsData.get_valid_sponsors)
-  }
-
-  useEffect(() => {
-    if (getValidSponsorError) {
-      setInternalError(true)
-    }
-  }, [getValidSponsorError])
-
-  useEffect(() => {
-    getSponsors()
-  }, [getValidSponsors])
-
-
-  const { error: getLifebankError, data: lifebankdata, refetch: getValidLifebank } = useQuery(GET_VALID_LIFEBANKS_QUERY, { fetchPolicy: 'cache-and-network' })
-
-  const getLifebanks = async () => {
-    await getValidLifebank()
-
-    if (lifebankdata && validLifebanks.length === 0) setValidLifebanks(lifebankdata.get_valid_lifebanks)
-  }
-
-  useEffect(() => {
-    if (getLifebankError) {
-      setInternalError(true)
-    }
-  }, [getLifebankError])
-
-
-  useEffect(() => {
-    getLifebanks()
-  }, [getValidLifebank])
 
   return (
     <BrowserRouter>
@@ -83,43 +36,6 @@ const App = ({ ual }) => {
             sidebarContent={<SideBar user={currentUser} onLogout={logout} triggerSideBarPosition={triggerSideBarPosition} />}
             sideBarPosition={sideBarPosition}
           >
-            {validSponsors.length > 0 && (
-              <>
-                {validSponsors.map((el, key) => (
-                  <LocalBusinessStructuredData
-                    key={key}
-                    name={el.name}
-                    openingHours={el.openingHours}
-                    address={el.address}
-                    logo={el.logo}
-                    email={el.email}
-                    location={el.location}
-                    telephone={el.telephone}
-                    socialMediaLinks={JSON.parse(el.social_media_links)}
-                  />
-                ))}
-              </>
-            )}
-            {validLifebanks.length > 0 && (
-              <>
-                {validLifebanks.map((element, key) => (
-                  <MedicalClinicStructuredData
-                    key={key}
-                    name={element.name}
-                    openingHours={element.openingHours}
-                    address={element.address}
-                    logo={element.logo}
-                    email={element.email}
-                    description={element.description}
-                    location={element.location}
-                    telephone={element.telephone}
-                  />
-                ))}
-              </>
-            )}
-            {internalError &&
-              <Redirect to="/internal-error" />
-            }
             <Grid container justify="center" alignItems="center">
               <Switch>
                 {routes.map(({ path, component: Component, ...args }) => (
